perf(render): SpectrogramImageProducer + FloatArray spectra [P1, P2, U2, plan 3.5]
- SpectrogramImageProducer owns the pixel pipeline: full-file renders
  downsampled to <=4096 columns (was one column per FFT frame, ~13k for
  5 min), computed on Dispatchers.Default via produceState, delivered as
  alternating double-buffered bitmaps — every data change repaints. The
  mutate-a-remembered-bitmap LaunchedEffect hack is gone [U2]: a loaded
  WAV renders immediately (closes the DEV-14 quirk).
- Spectra cross engine boundaries and live in histories as FloatArray
  (display precision, half the memory); DSP computation stays double —
  golden snapshot untouched.
- Paints hoisted out of the 43 Hz draw loop (filter bands, harmonic tags,
  manual tags, leader lines) [P2].

Emulator evidence (5-min 26 MB WAV): renders immediately with the correct
0..300 s axis; TOTAL PSS 242 MB / Java heap 161 MB — under the 256 MB
plan-§1 gate (old layout: 211 MB of double spectra + 52 MB bitmap).
73 tests green, lint 0, minified release builds.

+++ app/src/main/java/com/example/nvhspectro/LiveViewModel.kt
@@ -137,7 +137,7 @@ class LiveViewModel(application: Application, val session: MeasurementSession) :
 
         // Telemetry stays 1-to-1 with the audio display.
         val telemWithTtnr = telemetryForCalc.copy(
-            ttnrDb = (ttnrSpectrum.maxOrNull() ?: 0.0).toFloat(),
+            ttnrDb = ttnrSpectrum.maxOrNull() ?: 0f,
             trackedOrderDbFS = trackedDbFS,
             trackedOrderEmergenceDb = trackedEmergence
         )
